requirements.txt
langchain
python-dotenv
langchain_community
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
OpenAI